Revert "Feature/2 create list"
