+++ src/App.jsx
@@ -8,49 +8,47 @@ import CreateGroupPage from "./pages/CreatePages/CreateGroupPage";
 import TutorialPage from "./pages/CreatePages/TutorialPage";
 import CreateSpacePage from "./pages/CreatePages/CreateSpacePage";
 import GroupHomePage from "./pages/GroupHomePage";
+import NoGroupPage from "./pages/CreatePages/NoGroupPage";
 import GroupSpacePage from "./pages/GroupSpacePage";
-import NoGroupPage from "./pages/NoGroupPage";
 import { useAuthStatus } from "./hooks/useAuthStatus";
 import MyPage from "./pages/MyPage/MyPage";
-import GroupEvalPage from "./pages/GroupEvalPage";
 
 function App() {
-    const { isLoggedIn, hasGroup } = useAuthStatus();
+  const { isLoggedIn, hasGroup } = useAuthStatus();
 
-    return (
-        <Routes>
-            <Route path="/" element={<LoginPage />} />
-            <Route path="/signup" element={<SignupPage />} />
-            <Route path="/login" element={<LoginPage />} />
-            <Route path="/personality/1" element={<CleanPersonality_1 />} />
-            <Route path="/personality/2" element={<CleanPersonality_2 />} />
-            <Route path="/result" element={<CleanPersonality_3 />} />
-            <Route path="/noGroup" element={<NoGroupPage />} />
-            <Route path="/createGroup" element={<CreateGroupPage />} />
-            <Route path="/tutorial" element={<TutorialPage />} />
-            <Route path="/createSpace" element={<CreateSpacePage />} />
-            <Route path="/groupSpace" element={<GroupSpacePage />} />
-            <Route path="/groupHome" element={<GroupHomePage />} />
-            <Route path="/mypage" element={<MyPage />} />
-            <Route path="/groupEval" element={<GroupEvalPage />} />
+  return (
+    <Routes>
+      <Route path="/" element={<LoginPage />} />
+      <Route path="/signup" element={<SignupPage />} />
+      <Route path="/login" element={<LoginPage />} />
+      <Route path="/personality/1" element={<CleanPersonality_1 />} />
+      <Route path="/personality/2" element={<CleanPersonality_2 />} />
+      <Route path="/result" element={<CleanPersonality_3 />} />
+      <Route path="/noGroup" element={<NoGroupPage />} />
+      <Route path="/createGroup" element={<CreateGroupPage />} />
+      <Route path="/tutorial" element={<TutorialPage />} />
+      <Route path="/createSpace" element={<CreateSpacePage />} />
+      <Route path="/groupSpace" element={<GroupSpacePage />} />
+      <Route path="/groupHome" element={<GroupHomePage />} />
+      <Route path="/mypage" element={<MyPage />} />      
 
-            <Route
-                path="/redirect"
-                element={
-                    isLoggedIn ? (
-                        hasGroup ? (
-                            <Navigate to="/createSpace" replace />
-                        ) : (
-                            <Navigate to="/createGroup" replace />
-                        )
-                    ) : (
-                        <Navigate to="/" replace />
-                    )
-                }
-            />
-            <Route path="*" element={<div>잘못된 페이지입니다.</div>} />
-        </Routes>
-    );
+      <Route
+        path="/redirect"
+        element={
+          isLoggedIn ? (
+            hasGroup ? (
+              <Navigate to="/createSpace" replace />
+            ) : (
+              <Navigate to="/createGroup" replace />
+            )
+          ) : (
+            <Navigate to="/" replace />
+          )
+        }
+      />
+      <Route path="*" element={<div>잘못된 페이지입니다.</div>} />
+    </Routes>
+  );
 }
 
 export default App;
