+++ src/components/CheckList/GListAddModal.jsx
@@ -1,14 +1,14 @@
 import "./GListAddModal.css";
 import "react-datepicker/dist/react-datepicker.css";
 import { useContext, useState } from "react";
-import Modal from "../../../Modal";
-import Button from "../../../Button";
+import Modal from "../Modal";
+import Button from "../Button";
 
 import {
     toCleanStateContext,
     toCleanDispatchContext,
-} from "../../../../context/GroupContext";
-import { getBadgeImage } from "../../../../utils/get-badge-images";
+} from "../../pages/GroupSpacePage";
+import { getBadgeImage } from "../../utils/get-badge-images";
 import DatePicker from "react-datepicker";
 
 import { ko } from "date-fns/locale";
@@ -23,7 +23,6 @@ const GListAddModal = ({ isAddMode, setIsAddMode, selectedPlace }) => {
     const [selectedDate, setSelectedDate] = useState(null);
     const [createData, setCreateData] = useState({
         target: "group",
-        parentPlace: "none",
         place: selectedPlace,
         toClean: "",
         deadLine: "미정",
@@ -46,7 +45,6 @@ const GListAddModal = ({ isAddMode, setIsAddMode, selectedPlace }) => {
             createData.target,
             createData.name,
             createData.badgeId,
-            createData.parentPlace,
             createData.place,
             createData.toClean,
             createData.deadLine
