--- src/components/GroupSpace/NeedClean/NeedClean.css
@@ -1,17 +1,0 @@
-.NeedClean {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.NeedClean > h3 {
-    font-size: 22px;
-    font-weight: bold;
-    margin: 25px;
-}
-
-.NeedClean .NCleanItem_wrapper {
-    display: flex;
-    gap: 30px;
-}
