+++ src/components/Sidebar.css
@@ -1,85 +1,85 @@
 @font-face {
-    font-family: "NotoSansKR-Regular";
-    src: url("/fonts/NotoSansKR-Regular.ttf") format("truetype");
-    font-weight: normal;
-    font-style: normal;
+  font-family: "NotoSansKR-Regular";
+  src: url("/fonts/NotoSansKR-Regular.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
 }
 
 .sidebar {
-    width: 11.5vw;
-    max-width: 200px;
-    min-width: 150px;
-    min-height: 100vh;
+  width: 11.5vw;
+  max-width: 200px;
+  min-width: 150px;
+  min-height: 100vh;
 
-    display: flex;
-    flex-direction: column;
-    padding: 0;
-    box-sizing: border-box;
-    background-color: #ffffff;
+  display: flex;
+  flex-direction: column;
+  padding: 0;
+  box-sizing: border-box;
+  background-color: #ffffff;
 
-    font-family: "NotoSansKR-Regular", sans-serif;
+  font-family: "NotoSansKR-Regular", sans-serif;
 }
 
 .sidebar-home-inner {
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    gap: 5px;
-    padding: 4px 0;
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  gap: 5px;
+  padding: 4px 0;
 }
 
 .sidebar-home {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 8.5px 0;
-    font-size: 1rem;
-    font-weight: 500;
-    color: #1a1a1a;
-    font-family: "NotoSansKR-Regular", sans-serif;
-    background-color: #e2fdef;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 8.5px 0;
+  font-size: 1rem;
+  font-weight: 500;
+  color: #1a1a1a;
+  font-family: "NotoSansKR-Regular", sans-serif;
+  background-color: #e2fdef;
 }
 
 .sidebar-home-icon {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .sidebar-home-title {
-    font-weight: 600;
+  font-weight: 600;
 }
 
 /* 공간 편집하기 버튼 */
 
 .sidebar-edit-btn {
-    display: block;
-    background: #8be2b6;
-    border: none;
+  display: block;
+  background: #8be2b6;
+  border: none;
 
-    padding: 8.5px 0;
-    font-size: 1rem;
-    font-weight: 500;
-    color: #1a1a1a;
-    text-align: center;
+  padding: 8.5px 0;
+  font-size: 1rem;
+  font-weight: 500;
+  color: #1a1a1a;
+  text-align: center;
 }
 
 /* 공간 리스트 섹션 */
 
 .sidebar-section {
-    margin-bottom: 0;
+  margin-bottom: 0;
 }
 
 .sidebar-section-title {
-    display: block;
-    background: #ededed;
-    /* border: 1px solid #f0f0f0; */
+  display: block;
+  background: #ededed;
+  /* border: 1px solid #f0f0f0; */
 
-    padding: 8.5px 0;
-    font-size: 1rem;
-    font-weight: 500;
-    color: #1a1a1a;
-    text-align: center;
+  padding: 8.5px 0;
+  font-size: 1rem;
+  font-weight: 500;
+  color: #1a1a1a;
+  text-align: center;
 }
 
 /* 
@@ -90,16 +90,16 @@
 } */
 
 .sidebar-list-item {
-    background: #ffffff;
-    border: 1px solid #ededed;
-    border-bottom: none; /* 마지막 아이템 제외하고 아래쪽만 없애기 */
-    padding: 7.5px 0;
-    font-size: 0.9rem;
-    color: #1a1a1a;
-    cursor: pointer;
+  background: #ffffff;
+  border: 1px solid #ededed;
+  border-bottom: none; /* 마지막 아이템 제외하고 아래쪽만 없애기 */
+  padding: 7.5px 0;
+  font-size: 0.9rem;
+  color: #1a1a1a;
+  cursor: pointer;
 
-    text-align: center;
-    transition: background 0.15s;
+  text-align: center;
+  transition: background 0.15s;
 }
 /* 
 .sidebar-list-item:first-child {
@@ -108,16 +108,16 @@
 } */
 
 .sidebar-list-item:last-child {
-    border-bottom: 1px solid #ededed;
+  border-bottom: 1px solid #ededed;
 }
 
 .sidebar-list {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    overflow: hidden;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  overflow: hidden;
 }
 
 .sidebar-list-item:hover {
-    background: #eaf7f1;
+  background: #eaf7f1;
 }
