--- src/components/groupHome/GEvalItem.css
@@ -1,80 +1,0 @@
-.GEvalItem {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-/* badge 이미지 & 이름*/
-.GEvalItem > img {
-    width: clamp(77.8px, 7.71vw, 111px);
-    height: clamp(77.8px, 7.71vw, 111px);
-    margin-top: clamp(28px, 2.78vw, 40px);
-    border-radius: 50%;
-    border: clamp(4.26px, 0.416vw, 6px) solid rgb(164, 235, 217);
-}
-
-.GEvalItem .name_text {
-    margin-top: clamp(7.7px, 0.76vw, 11px);
-    font-size: clamp(16.8px, 1.67vw, 24px);
-}
-
-/* 청소완료 개수 */
-.GEvalItem .done_section {
-    width: clamp(93px, 8.9vw, 133px);
-    height: clamp(29.4px, 2.8vw, 42px);
-    border-radius: 100px;
-    margin-top: clamp(7.7px, 0.76vw, 11px);
-
-    background-color: #e2fdef;
-
-    display: flex;
-    gap: clamp(10.5px, 1.46vw, 15px);
-    justify-content: center;
-    align-items: center;
-}
-
-.GEvalItem .done_text {
-    font-size: clamp(8.4px, 1.11vw, 12px);
-}
-
-.GEvalItem .done_count {
-    font-size: clamp(16.8px, 1.67vw, 24px);
-    font-family: bold;
-    color: #1ac2a9;
-    margin-bottom: clamp(1.4px, 0.28vw, 2px);
-}
-
-/* 청소 유형 평가 */
-.GEvalItem .cleanPersonality_section {
-    display: flex;
-    gap: clamp(5.6px, 0.56vw, 8px);
-    justify-content: center;
-    align-items: center;
-
-    margin-top: clamp(14px, 1.39vw, 20px);
-}
-
-.GEvalItem .cleanPersonality_section .cleanPersonality_type {
-    font-size: clamp(11.2px, 1.11vw, 16px);
-    font-family: bold;
-}
-
-.GEvalItem .cleanPersonality_section .cleanPersonality_dis {
-    font-size: clamp(9.8px, 0.97vw, 14px);
-}
-
-/* 별점 선택 */
-.GEvalItem .rating {
-    margin-top: clamp(15.4px, 1.53vw, 22px);
-}
-
-.GEvalItem .rating .star {
-    background-color: transparent;
-    border: none;
-}
-
-.GEvalItem .rating .star > img {
-    width: clamp(21px, 2.08vw, 30px);
-    height: clamp(21px, 2.08vw, 30px);
-}
