--- src/components/groupHome/GroupEval.css
@@ -1,79 +1,0 @@
-.GroupEval {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-/* 상단 제목 및 날짜 */
-.GroupEval > h3 {
-    margin-top: clamp(42px, 4.1vw, 59px);
-    font-size: clamp(20.16px, 2vw, 28px);
-    font-weight: bold;
-}
-
-.GroupEval .date {
-    margin-top: clamp(2.1px, 0.29vw, 3px);
-    font-size: clamp(15.12px, 1.46vw, 21px);
-    color: #787878;
-}
-
-/* 하단에 평가하는 부분 */
-.GroupEval .eval_section {
-    display: flex;
-    gap: clamp(28px, 4.17vw, 40px);
-    margin-top: clamp(35px, 4.86vw, 50px);
-}
-
-/* 자기 평가 */
-.GroupEval .me {
-    width: clamp(238.22px, 23.26vw, 335px);
-    height: clamp(357px, 34.86vw, 502px);
-    background-color: white;
-    border-radius: clamp(14px, 1.39vw, 20px);
-}
-
-.GroupEval .me > h3 {
-    font-size: clamp(15.84px, 1.53vw, 22px);
-    font-family: bold;
-    text-align: center;
-    margin-top: clamp(35.5px, 3.47vw, 50px);
-    margin-bottom: 0px;
-}
-
-/* 그룹원 평가 */
-.GroupEval .other {
-    width: clamp(657.7px, 64.23vw, 925px);
-    height: clamp(357px, 34.86vw, 502px);
-    background-color: white;
-    border-radius: clamp(14px, 1.39vw, 20px);
-}
-
-.GroupEval .other > h3 {
-    font-size: clamp(15.64px, 1.53vw, 22px);
-    font-family: bold;
-    text-align: center;
-    margin-top: clamp(35.5px, 3.47vw, 50px);
-    margin-bottom: 0px;
-}
-
-.GroupEval .other .otherItem {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: clamp(56.8px, 5.5vw, 80px);
-}
-
-/* 저장 버튼 */
-.GroupEval .Button_save {
-    width: clamp(88px, 8.75vw, 126px);
-    height: clamp(40px, 4.03vw, 58px);
-    border-radius: clamp(10.5px, 1.04vw, 15px);
-    color: white;
-    margin: clamp(21px, 2.08vw, 30px) 0px clamp(14.2px, 1.39vw, 20px)
-        clamp(836.97px, 81.73vw, 1177px);
-    font-size: clamp(20.16px, 2vw, 28px);
-    font-family: bold;
-    text-align: center;
-    line-height: 0.5px;
-}
