--- src/components/groupHome/GroupHome.css
@@ -1,124 +1,0 @@
-.GroupHome {
-    margin: clamp(28px, 2.78vw, 40px) clamp(50px, 4.86vw, 70px);
-    background-color: rgb(245, 245, 245);
-}
-
-/* 상단 부분*/
-.GroupHome .groupName {
-    display: flex;
-    gap: clamp(7px, 0.69vw, 10px);
-    justify-content: center;
-    align-items: center;
-    margin-bottom: clamp(28px, 2.78vw, 40px);
-}
-
-.GroupHome h3 {
-    font-size: clamp(22px, 2.22vw, 32px);
-    font-weight: bold;
-}
-
-.GroupHome .groupName .pencil_img {
-    width: clamp(20px, 2.08vw, 30px);
-    height: clamp(20px, 2.08vw, 30px);
-    margin-bottom: clamp(4px, 0.42vw, 6px);
-}
-
-.GroupHome .groupName .home_img {
-    width: clamp(23px, 2.43vw, 35px);
-    height: clamp(18px, 1.88vw, 27px);
-    margin-bottom: clamp(5px, 0.56vw, 8px);
-}
-
-/* 하단 부분 */
-.GroupHome .groupHomeUnder {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: clamp(27px, 2.78vw, 40px);
-}
-
-/* 공간 구조도 */
-.GroupHome .groupSpace {
-    width: clamp(494px, 48.33vw, 696px);
-    height: clamp(494px, 48.33vw, 696px);
-    background-color: white;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: clamp(14px, 1.39vw, 20px);
-    margin-bottom: clamp(28px, 2.78vw, 40px);
-}
-
-/* 그룹 규칙 & 평가 */
-.GroupHome .groupHomeRE {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: clamp(20px, 2.08vw, 30px);
-    margin-bottom: clamp(28px, 2.78vw, 40px);
-}
-
-/* 그룹 규칙 */
-.GroupHome .groupRule {
-    width: clamp(386px, 39.17vw, 564px);
-    height: clamp(187.7px, 18.33vw, 264px);
-    background-color: white;
-    border-radius: clamp(14px, 1.39vw, 20px);
-    padding: clamp(28px, 3.05vw, 40px) clamp(38px, 3.82vw, 55px);
-}
-
-.GroupHome .groupRule > h3 {
-    text-align: center;
-    font-size: clamp(16px, 1.6vw, 23px);
-    font-weight: bold;
-    margin-bottom: clamp(21px, 2.08vw, 30px);
-}
-
-.GroupHome .groupRule .ruleContent {
-    font-size: clamp(12px, 1.18vw, 17px);
-    white-space: pre-line;
-    line-height: clamp(30px, 3.05vw, 44px);
-}
-
-/* 그룹 평가 */
-.GroupHome .groupEval {
-    width: clamp(386px, 39.17vw, 564px);
-    height: clamp(285.86px, 27.9vw, 402px);
-    background-color: white;
-    border-radius: clamp(14px, 1.39vw, 20px);
-    padding: clamp(28px, 3.05vw, 40px);
-    position: relative;
-}
-
-.GroupHome .groupEval > h3 {
-    text-align: center;
-    font-size: clamp(16px, 1.6vw, 23px);
-    font-weight: bold;
-    margin-bottom: clamp(38px, 3.82vw, 55px);
-}
-
-.GroupHome .Button_eval {
-    position: absolute;
-    top: clamp(10px, 1.04vw, 15px);
-    right: clamp(10px, 1.04vw, 15px);
-    width: clamp(91px, 9.03vw, 130px);
-    height: clamp(28px, 2.78vw, 40px);
-    border-radius: clamp(10px, 1.04vw, 15px);
-    color: white;
-    font-size: clamp(11px, 1.04vw, 15px);
-}
-
-.GroupHome .GInfoItems {
-    width: clamp(337px, 33.05vw, 476px);
-    height: clamp(163px, 16.04vw, 231px);
-    display: flex;
-    flex-direction: column;
-    gap: clamp(48px, 4.79vw, 69px);
-}
-
-.GroupHome .EvalRow1,
-.GroupHome .EvalRow2 {
-    display: flex;
-    gap: clamp(33px, 3.33vw, 48px);
-}
