--- src/components/groupHome/TListItem.css
@@ -1,52 +1,0 @@
-.TListItem {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: clamp(14.08px, 2vw, 28.8px);
-    width: clamp(361.6px, 35.28vw, 508px);
-    height: clamp(54.54px, 5.4vw, 77.76px);
-    border-radius: clamp(10.5px, 1.04vw, 15px);
-    background-color: rgb(248, 248, 248);
-    margin-bottom: clamp(6.36px, 0.63vw, 9.072px);
-    position: relative;
-}
-
-.TListItem .target {
-    position: absolute;
-    left: clamp(14.22px, 1.39vw, 20px);
-    font-size: clamp(10.56px, 1.1vw, 15.84px);
-}
-
-.TListItem .place {
-    width: clamp(33.99px, 3.32vw, 47.8px);
-    font-size: clamp(11.26px, 1.1vw, 15.84px);
-    text-align: center;
-    position: absolute;
-    left: clamp(64px, 6.25vw, 90px);
-}
-
-.TListItem .toClean {
-    font-size: clamp(10.56px, 1.1vw, 15.84px);
-    width: clamp(78.6px, 11vw, 158.4px);
-    position: absolute;
-    left: clamp(123.02px, 12.01vw, 173px);
-}
-
-.TListItem .deadLine {
-    width: clamp(30.93px, 3.02vw, 43.5px);
-    text-align: center;
-    font-size: clamp(10.56px, 1.1vw, 15.84px);
-    position: absolute;
-    left: clamp(261.69px, 25.56vw, 368px);
-}
-
-.TListItem .Button_done {
-    width: clamp(35.7px, 3.47vw, 49.968px);
-    height: clamp(20.16px, 2vw, 28.8px);
-    border-radius: clamp(7.14px, 0.69vw, 10px);
-    position: absolute;
-    right: clamp(14.22px, 1.39vw, 20px);
-    color: white;
-    font-size: clamp(9.26px, 0.9vw, 12.96px);
-    font-weight: bold;
-}
