--- src/context/GroupContext.js
@@ -1,5 +1,0 @@
-// context/GroupContext.js
-import { createContext } from "react";
-
-export const toCleanStateContext = createContext();
-export const toCleanDispatchContext = createContext();
