+++ src/index.css
@@ -22,5 +22,4 @@ body {
 html,
 #root {
     background-color: rgb(245, 245, 245);
-    height: 100%;
 }
