--- src/pages/CreateGroupPage.css
@@ -1,201 +1,0 @@
-@font-face {
-  font-family: "NotoSansKR-Regular";
-  src: url("/fonts/NotoSansKR-Regular.ttf") format("truetype");
-  font-weight: 400;
-  font-style: normal;
-}
-@font-face {
-  font-family: "NotoSansKR-Bold";
-  src: url("/fonts/NotoSansKR-Bold.ttf") format("truetype");
-  font-weight: 700;
-  font-style: normal;
-}
-
-body {
-  font-family: "NotoSansKR-Regular", sans-serif;
-  font-weight: 400;
-}
-
-.create-group-bg {
-  background-color: #f5f5f5;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.create-group-card {
-  width: 25vw;
-  min-width: 350px;
-  margin: 50px auto;
-  background: #ffffff;
-  border-radius: 20px;
-  box-shadow: 0 2px 8px #d9d9d9;
-  padding: 20px 25px 30px 25px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.create-group-title {
-  font-family: "NotoSansKR-Bold", sans-serif;
-  font-weight: 700;
-  font-size: 1.8rem;
-  margin-bottom: 30px;
-  text-align: center;
-}
-
-.create-group-form {
-  width: 90%;
-}
-
-.form-section {
-  margin-bottom: 25px;
-}
-
-.form-label {
-  font-family: "NotoSansKR-Regular", sans-serif;
-  font-weight: 400;
-  font-size: 1rem;
-}
-
-.form-name {
-  width: 100%;
-  margin-top: 10px;
-  height: 48px;
-  padding: 0 20px;
-  border-radius: 15px;
-  border: none;
-  background-color: #f5f5f5;
-  font-size: 0.9rem;
-  outline: none;
-  box-sizing: border-box;
-  font-family: "NotoSansKR-Regular", sans-serif;
-  vertical-align: middle;
-}
-
-.form-email {
-  width: 100%;
-  height: 3rem;
-  padding: 0 20px;
-  border-radius: 15px;
-  border: none;
-  background-color: #f5f5f5;
-  font-size: 0.9rem;
-  outline: none;
-  box-sizing: border-box;
-  font-family: "NotoSansKR-Regular", sans-serif;
-  vertical-align: middle;
-}
-
-.form-rule {
-  width: 100%;
-  margin-top: 10px;
-  padding: 15px 20px;
-  border-radius: 15px;
-  border: none;
-  background-color: #f5f5f5;
-  font-size: 0.9rem;
-  resize: none;
-  outline: none;
-  box-sizing: border-box;
-  font-family: "NotoSansKR-Regular", sans-serif;
-}
-
-/* 이메일 입력 ~ 초대 버튼 줄 */
-.member-input-row {
-  display: flex;
-  gap: 15px;
-  margin-top: 10px;
-}
-
-.invite-btn {
-  background: #8be2b6;
-  color: #ffffff;
-  border: none;
-  border-radius: 15px;
-  padding: 0 20px;
-  font-family: "NotoSansKR-Bold", sans-serif;
-  font-weight: 400;
-  font-size: 0.9rem;
-  cursor: pointer;
-  transition: background 0.2s;
-  height: 3rem;
-  box-sizing: border-box;
-  white-space: nowrap;
-}
-.invite-btn:hover {
-  background: #74d3a4;
-}
-
-.member-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 15px;
-  margin-top: 12px;
-}
-
-.member-chip {
-  display: flex;
-  align-items: center;
-  background: #ffffff;
-  border: 1.5px solid #ececec;
-  border-radius: 20px;
-  padding: 3px 15px 5px 15px;
-  font-size: 0.9rem;
-  font-family: "NotoSansKR-Regular", sans-serif;
-  font-weight: 400;
-  box-shadow: 0 1px 1.5px #d9d9d9;
-}
-
-.remove-member-btn {
-  margin-left: 10px;
-  background: none;
-  border: none;
-  color: #676767;
-  font-weight: 400;
-  font-size: 0.9rem;
-  cursor: pointer;
-  padding: 0;
-  font-family: "NotoSansKR-Regular", sans-serif;
-}
-.remove-member-btn:hover {
-  color: #74d3a4;
-}
-
-.create-btn {
-  width: 100%;
-  background: #8be2b6;
-  color: #ffffff;
-  border: none;
-  border-radius: 15px;
-  padding: 10px 0;
-  font-family: "NotoSansKR-Bold", sans-serif;
-  font-weight: 700;
-  font-size: 1.4rem;
-  margin-top: 15px;
-  cursor: pointer;
-  transition: background 0.2s;
-}
-.create-btn:hover {
-  background: #74d3a4;
-}
-
-/* 스크롤 */
-.form-rule::-webkit-scrollbar {
-  width: 7.5px;
-}
-
-.form-rule::-webkit-scrollbar-track {
-  background: #f1f1f1;
-  border-radius: 15px;
-}
-
-.form-rule::-webkit-scrollbar-thumb {
-  background: #c1c1c1;
-  border-radius: 15px;
-  transition: background 0.3s ease;
-}
-
-.form-rule::-webkit-scrollbar-thumb:hover {
-  background: #a8a8a8;
-}
