--- src/pages/CreateSpacePage.jsx
@@ -1,7 +1,0 @@
-import Header from "../components/Header";
-
-function CreateSpacePage() {
-  return <h1>Create Space Page</h1>;
-}
-
-export default CreateSpacePage;
