--- src/pages/GroupEvalPage.jsx
@@ -1,16 +1,0 @@
-import Header from "../components/Header";
-import GroupEval from "../components/groupHome/GroupEval";
-import GroupProvider from "../context/GroupProvider";
-
-const GroupEvalPage = () => {
-    return (
-        <div>
-            <GroupProvider>
-                <Header />
-                <GroupEval />
-            </GroupProvider>
-        </div>
-    );
-};
-
-export default GroupEvalPage;
