--- src/pages/TutorialPage.jsx
@@ -1,7 +1,0 @@
-import Header from "../components/Header";
-
-function TutorialPage() {
-  return <h1>Tutorial Page</h1>;
-}
-
-export default TutorialPage;
